[FIX] account_reports: test_tax_report_grid can fail at midnight
The test creates an invoice, then a reversal, then checks the report
values on the day of the invoice.
But currently, there is no date for the reversal, so if the invoice
is created at 23:59:59 and the reversal at 00:00:01, the reversal
won't be in the report values and so, the test will fail.

The solution proposed is just to create the reversal at the invoice date

Linked to runbot error 59289

closes odoo/enterprise#59302

X-original-commit: ae71aee1207ffd9c54d84d647110383e1d5835c6
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>