[FIX] web_enterprise: tooling: expose owl module for auto-import
On an IDE (VSCode) with a TypeScript Server.
For some reason, it was not possible to have imports suggestion to OWL in enterprise after commit [1].

Our main hypothesis is that, due to commit [1] who points to odoo/addons, import lines in those files
are resolved first in the scope of the odoo directory.

Consider a file in enteprise:
```js
import { registry } from "@web/core";
import { Component } from "@odoo/owl"

```

From the TSServer point of view it needs to resolve `@web` first, which itself has to resolve `@odoo/owl` in its directory scope,
that is, in the scope of odoo community. It then keeps it in cache for later and importantly shadows the one present in enterprise/node_modules.
Only then the explicit `@odoo/owl` is resolved with the odoo community's owl module from the cache.

Since that owl module from the cache is not explicitly available in enterprise, there are no imports suggestions.

This commit defines a main.d.ts file that only exposes the community's `@odoo/owl` explicitly.
This hypothesis is confirmed by some hints that we discover during debugging:
- delete odoo/nodule_modules in enterprise: @odoo/owl becomes a suggestion again.
- comment the path line of @web/* in enterprise/jsconfig.json, owl is suggested.
- replace the copy of node_modules in enterprise by a symlink to odoo/node_modules, owl is suggested.
This last one should also be considered as a valid fix for the current issue, as node_modules in odoo and enterprise *are the same*.

After this commit, owl is proposed for imports in VSCode.

TypeScript issue: https://github.com/microsoft/TypeScript/issues/57762

[1]: 62cbb20

closes odoo/enterprise#58757

X-original-commit: 99082e2434efef67bf350950226dd09de5c79dd2
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>