[FIX] l10n_be_codabox*: prevent non-Accounting Firms from using the module
We want to prevent non-Accounting Firms from using this module since
it may generate some tickets on our side, and on CodaBox's side.

From now on, if no Accounting Firm is set on the Company, we:
- hide the CodaBox connection section in the Settings
- raise an error when installing the module

If a DB has not yet the CodaBox module installed and no Accounting Firm
is set, we will raise an error when the user tries to install the module

If a DB has already the CodaBox module installed, no Accounting Firm,
and the module has not been upgraded yet after this commit, we will
raise an error when the user tries to click on Manage Connection

If a DB has already the CodaBox module installed, no Accounting Firm,
and the module has been upgraded after this commit, the CodaBox
Connection section is simply hidden.

A demo Accounting Firm has been added for Belgium to avoid
an error when installing the CodaBox module when running all_l10n test.

The sections were also renamed.

task-id 3795251

closes odoo/enterprise#58787

X-original-commit: 280827c
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>