[FIX] web_studio: skip studio_approval_rule patch on non-existing models
If a studio.approval.rule refers to a non-existing model (e.g. the model was
removed, or renamed, or it's a custom model that is not present during an
upgrade), then while attempting to patch the rule's method the registry will
return a None model, which then causes an AttributeError.

We add a check for the Model's existence, so that any None models are skipped.
This follows the approach in base_automation, which was the technical
inspiration for the commit that implemented the approval rule logic.

Ticket: 3754939, 3777251
```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/17.0/odoo/service/server.py", line 1302, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "<decorator-gen-16>", line 2, in new
  File "/home/odoo/src/odoo/17.0/odoo/tools/func.py", line 87, in locked
    return func(inst, *args, **kwargs)
  File "/home/odoo/src/odoo/17.0/odoo/modules/registry.py", line 113, in new
    odoo.modules.load_modules(registry, force_demo, status, update_module)
  File "/home/odoo/src/odoo/17.0/odoo/modules/loading.py", line 606, in load_modules
    model._register_hook()
  File "/home/odoo/src/enterprise/17.0/web_studio/models/studio_approval.py", line 203, in _register_hook
    _patch(Model, approval.method, approval_method)
  File "/home/odoo/src/enterprise/17.0/web_studio/models/studio_approval.py", line 132, in _patch
    ModelClass = model.env.registry[model._name]
AttributeError: 'NoneType' object has no attribute 'env'
```

closes odoo/enterprise#58688

Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>