[FIX] l10n_be_hr_payroll: only count active employees
As we are getting the employees based on the leaves, and the leaves are
not archived when an employee is archived, the counter will count the
archived employees.

When we display the list of employees with the action generated on the
warning an active_test=True is used by default and so the count doesn't
correspond to the amount of record displayed.

To avoid such issue, we only add the active employees

closes odoo/enterprise#58671

X-original-commit: 957f5cf2f19da8a1e51610ce1cb71d6dc1a878ea
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>