[PERF] spreadsheet: return errors values instead of throwing
At the initial evaluation of a spreadsheet with pivots/lists, almost all cells
are in error because they are loading. But throwing errors is slow.

With this commit, instead of throwing loading errors for every single cell,
the error is returned.

This commit improves the first evaluation time of spreadsheet the
"BE Timesheet" dashboard on odoo.com by ~20%, from ~1716ms to ~1365ms

closes odoo/enterprise#58643

Task: 3736190
X-original-commit: 904349dee97539e6579c4a65315f0408a79e79b2
Related: odoo/odoo#157654
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>