[FIX] documents: file extension on replacing document
Steps to reproduce:
- Enable debug mode.
- Select a document (say a pdf file).
- Replace it with a non-pdf file (say png).

Issue:
Currently, the file extension is not updated on replacing a document
according to the newly uploaded document.

Cause:
The compute method of file extension fetched and set its value only if it was
not set already.  As a result, the extension was never updated as the field
already had a set value (of older file).

Fix:
This commit alters the code of `_compute_file_extension` to ensure that the
extension is recomputed and set every time the name of a document file changes.

Task-3801686

closes odoo/enterprise#60618

X-original-commit: 65fe7989a3385fa722b638fcdb38d1382c08c952
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Khushi Patel (khpa) <khpa@odoo.com>