[FIX] planning: added filter groupby
Before this commit there was no groupby Properties.

In this commit we add a filter for groupby properties in planning module.
The search view is being used by two modules`planning.slot` and
`planning.analysis.report` thus solution resulted in adding the filter in
`planning_view_search` as it is a primary inherited view of search view base in
planning.slot.

task-3801551

closes odoo/enterprise#58814

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>