[FIX] l10n_lu_hr_payroll: add back xml ids
In https://github.com/odoo/enterprise/pull/58054
by mistake one record has been removed, and one xml
id has been changed. Which is errorious for stable.
To solve the issue, we add back records.
We added back l10n_lu_employees_atn_transport, in
case it is referenced, but we make sure that it won't be
added in in the payslip computation.

closes odoo/enterprise#58658

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>