[FIX] sale_subscription: unexpected discount recomputation
`discount` field on `sale.order.line` model is configured in sale_subscription
to be recomputed when the order `subscription_state` is modified.

Therefore, updates to that field should be avoided unless necessary.

Nevertheless, in the override of `action_confirm`, the subscription state
was always updated to False for non subscription orders, leading to an
unexpected recomputation of discounts.

opw-3740645

closes odoo/enterprise#58673

Related: odoo/odoo#157699
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>