[FIX] account_reports: error formating value GL
In General Ledger report, we get an error when trying
to reformat value (K$, M$, ..)

Reason:

When setting the 'currency' value on each column, we take the
currency of the account,  which is interpreted as a string on
js side, then used to call format_value in where we try to
get `currency.id`, leading to the error.
To avoid that, we return only the id of the currency if there is one.

Also, we get a KeyError whendoing the same flow with unfolded lines

opw-3793209

closes odoo/enterprise#58678

Signed-off-by: William André (wan) <wan@odoo.com>