[FIX] planning: fix recurrence when we have overlapping slots
When generating recurrent shifts, we check whether the resource
is busy or not and if they are, we generate the recurrent shift
as an open shift. Currently, we check whether the resource is
busy by seeing whether there are any overlapping slots already
for that resource and by checking that the allocating percentage
for all overlapping slots wasn't above 100%. 

This commit changes the implementation in order for it to work
with different duration slots. We do this by comparing the total
planned hours between the overlapping slots to the total hours in
the overlapping period.

**Disclaimer:** This does not respect the working hours of the employee:
Let Bob work from 9:00 to 17:00. If shift A is planned from 8:00
to 17:00 (8h) and recurrent shift B is planned from 16:00 to 17:00 (1h).
In this scenario, the recurrent shift will be planned normally because
the total hours in the overlapping period are 9 and the total hours being
planned are also 9.

task-3806681

closes odoo/enterprise#58693

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>