[FIX] documents: close preview when documents displayed change
How to reproduce:
- Go to documents
- Click on a workspace with documents
- Click on a document to preview it
- Click on another workspace/select a tags/select a tag group, change the
filter/domain

The preview is still displayed but shouldn't as we can't see the content of the
newly selected/filtered. This fix ensures that the preview is closed in such
circumstances.

Technical notes:
- the event was already trigerred but on the wrong bus. This fix triggers it on
the right one.
- we update the test "document inspector: edit without selecting" as it alters
the tags simply by opening its preview and the change made here cause the
preview to be closed automatically when tags are modified. Therefore, we add
the reopening of the preview right after each tag modification to prevent
altering the original test's purpose.

Task-3624235

closes odoo/enterprise#61245

X-original-commit: 74eb051871df3066a6eaeabf768658cbe2f9b901
Signed-off-by: Pierre-Yves Dufays (pydu) <pydu@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>