[IMP] *: use path-based routing for client actions
Since [1], path-based routing is supported for window actions, and
server actions. The support for client actions was incomplete. It wasn't
possible to add a path if the client action wasn't on the database;
at the reload, the client actions didn't have a proper breadcrumb; and
client actions didn't have the possibility of using resId (only have
action_id as props).

This commit, fix this issues by giving the possibility to add a path,
and a display name to the client action in JS. Also, it adds resId to
the props given to the client action as well as the updateResId
function, which allows the client action to update the resId on the URL
as a form view will do.

[1] : https://github.com/odoo/odoo/commit/c63d14a0485a553b74a8457aee158384e9ae6d3f

closes odoo/enterprise#58765

Related: odoo/odoo#157867
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>