[FIX] account_intrastat: Format column correctly
When a user tries to print in PDF the intrastat report,
he gets a traceback because the pdf template tries to call
`len()` on the name and the name equals `None`.
The aim of this commit is using the `_build_column_dict` method
instead of formatting the column by ourselves.
By doing this, the `None` value are set for an empty string.

no task id

closes odoo/enterprise#58793

X-original-commit: 193878c10dee796e007e9a7515c3fcf42f59b2f4
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>