[REF] web_studio: addStructure separation of concerns
The `addStructure` method was doing too much, it implemented the logic
of adding a structure for all the views editor.

This commit refactor `addStructure` method to have a better separation
of concerns, the logic of adding a structure can now be implemented
specificallyin the `addViewStructure` function of each view editor.

closes odoo/enterprise#64852

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>