[FIX] delivery_fedex: Fix package commodity CustomsValue
The Customs values for a commodity needs to be the values
of the entire package. Currently, the unit value is set which
results in the following warning.

```
{
    'Severity': 'WARNING',
    'Source': 'crs',
    'Code': '448',
    'Message': 'The sum of internationalDetail commodities customs value
        amounts do not the equal the internationalDetail customs value
        amount; the greater customs value amount was used to rate.',
    'LocalizedMessage': 'The sum of internationalDetail commodities
        customs value amounts do not the equal the internationalDetail
        customs value amount; the greater customs value amount was used
        to rate.',
    'MessageParameters': []
}
```

previous fix https://github.com/odoo/enterprise/pull/38056 missed to
correct the value for rating api so a following fix
https://github.com/odoo/odoo/pull/116068 was made, which made the first
fix obsolete + incorrect. The followup PR missed adjusting the FedEx
value to match its change (i.e. other carriers were checked, but FexEx
was overlooked)
This commit corrects this issue and sets the correct commodity value for
FedEx.

X-original-commit: 5be067f81ab9f013e4531bda9bdfee6ff64d73d0
Part-of: odoo/enterprise#59052