[FIX] appointment: show conflict for new events
In the gantt view the "conflict" warning when a user is already booked
for another appointment does not show up before the appointment is saved.

The fix is to get the actual id of the partner instead of "newid"
that's always set in new onchange calls.

task-3700412

Part-of: odoo/enterprise#58245