[FIX] planning_gannt_renderer: fix flexible hours passed undefined value
Issue:
isFlexibleHours is expecting a resource's id to find whether the resource's working calendar is flexible or not. However, the argument passed to the function was not the resource id, but the resource itself. As the `!!` operator was used in the function, the returning value was always False (!!undefiend == False)

This commit fixes this by passing the id instead of the object.

related Task: 3762895

closes odoo/enterprise#58788

X-original-commit: 61afda686a786d4e44d82902c995c0265d742404
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>