[FIX] mrp_workorder_hr_account: modify account if WO time deleted
Steps
-----
1. Install accounting and mrp_workorder_hr, activate analytic accounting in
settings.
2. Select a work center and check "Requires Log In"
3. Create a manufacturing order and add a work order using the work center from
the previous step. On the miscellaneous tab, add an analytic account.
4. Confirm the MO.
5. On the Work Orders tab, change the Real Duration to 60:00
6. Analytic account smart button > the gross margin is correctly set.
7. Either change the real duration to 0:00 or open the work order and delete a
line from the Time Tracking tab.
Note: decreasing the Real Duration on the Manufacturing Order page will not lead
to an issue as long as no time record is deleted on the work order (typically,
when decreasing the time by a small amount).

Issue
-----
The gross margin still takes into account the employee's cost for the deleted time.
Note that the workcenter's cost is removed from the gross margin as expected.
Either by _create_or_update_analytic_entry in case the time of the work order
is modified, or unlink if the work order is deleted.
https://github.com/odoo/odoo/blob/16.0/addons/mrp_account/models/mrp_workorder.py

Fix
-----
When deleting a time record, modify the analytic account accordingly.
We create an analytic entry with a previous_duration equal to the
double of the time that is deleted.
https://github.com/odoo/enterprise/blob/a992a3ae345478a148e536103c1585a4caa86113/mrp_workorder_hr_account/models/mrp_workorder.py#L16
That way, we create an analytic entry with a duration equal to the negative
of the deleted time. This will add a positive amount to the analytic
account.
https://github.com/odoo/enterprise/blob/a992a3ae345478a148e536103c1585a4caa86113/mrp_workorder_hr_account/models/mrp_workorder.py#L21-L22
This  makes the analytic account's gross margin consistent with the time
worked.

opw-3759144

closes odoo/enterprise#62794

X-original-commit: 28911bb86ab46931a9b456de1dba600967cb4dd1
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>