[FIX] l10n_mx_edi_pos: improve compute of `cfdi_to_public` on pos_order
How to reproduce:

- in point of sale, enable "Generate a code on ticket"
- create a pos_order, save the ticket
- using the code on the ticket, request an invoice through the portal
and fill the information as if you were a mexican company

=> An invoice is generated and sent successfully to the SAT, but "CFDI
to Public" is True on the invoice. Hence, the receptor's RFC will be
"XAXX..." instead of the RFC filled in the portal.

Reason:

The pos_order has "CFDI to Public" to True, and this value is copied on
the invoice.

Fix:

Recompute the value for the "CFDI to Public" field on the pos_order when
the partner_id is filled. It should only be True when the Receptor's RFC
is "XAXX...". Note that the new compute for "CFDI to Public" on the
pos_order now resembles the one on the account_move.

opw-3767576

closes odoo/enterprise#59012

X-original-commit: 39a0542b6505e18e3fdfbefba4210d8ba87bae3c
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>