[FIX] account_reports: xlsx export and option companies
The aim of this commit is to allow the user to make xslx export with
custom composite report.

Context:
The customer made a custom report in which he is calling some other
reports via the section feature.

Before this commit:
Exporting the report would result in a traceback because a keyError was raised
when trying to access the `companies` key which wasn't present in the
dictionnary.

What went wrong:
The initializers for the `companies` key weren't consistent with each other.
In some initializer, the `currency_id` key was set while it wasn't in other.
In the method `_add_options_xlsx_sheet` only the exact same options were kept
before pursuing the operations.
As the `currency_id` wasn't set in all the dict of the `options['companies']`
list, that key-value pair was ignored and wasn't passed down.

After this commit:
The file can be exported with success.

To solve the `currency_id` key difference issue and prevent it to ever happen
again, we centralize the place where the dictionnaries are created to be sure
the exact same keys are set for the `companies` options regardless of how those
companies are chosen.

Side notes:
- We aren't testing this case because it is a corner case in which the
  `filter_multi_company` has been updated to its parent value and then
  reset to their default value via an update of the module.
- We are deleting the `_multi_company_selector_init_options` method as
  this was the only place were it was used and it was never meant to be
  overriden.

opw-3802602

closes odoo/enterprise#59238

X-original-commit: 27a739a49b77aa96b96a24124a08a2c9fdef6f7b
Signed-off-by: John Laterre (jol) <jol@odoo.com>