[REF] timesheet_grid: rename "readonly"
`_get_timesheet_timer_data` return `readonly` if a timer is running in
another company. On the client side, we store this information in a
`headerReadonly` key.
Yet, this doesn't really speaks for itself, as, when this value is true,
the timer is not displayed, let alone readonly. What is displayed is a
message that says that a timer is running in another company.

So, with this commit, we rename both references respectively
`other_company` and `otherCompany`.

closes odoo/enterprise#56679

Related: odoo/odoo#154134
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>