[FIX] timesheet_grid: approvers can edit n-2 validated timesheets
In such a hierarchy:

Clara -> Maxime -> Gregory

Before this revision:

Maxime and Gregory can validate Clara's timesheets,
but only Maxime can edit fields of Clara's validated timesheets
thanks to the fact he is the direct manager of Clara.

Gregory is the manager of Clara by recursivity,
and the recursivity was correctly handled in the validation check,
but not in the edition of validated timesheets.

After this revision:

Maxime and Gregory can validate Clara's timesheets,
Maxime and Gregory can edit fields of Clara's validated
timesheets.

opw-3613473

closes odoo/enterprise#59296

X-original-commit: 741f99a4d62bd16750736b66c040f79f49b5ccb5
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>