[FIX] timesheet_grid: approvers can edit n-2 validated timesheets
In such a hierarchy:

Clara -> Maxime -> Gregory

Before this revision:

Maxime and Gregory can validate Clara's timesheets,
but only Maxime can edit fields of Clara's validated timesheets
thanks to the fact he is the direct manager of Clara.

Gregory is the manager of Clara by recursivity,
and the recursivity was correctly handled in the validation check,
but not in the edition of validated timesheets.

After this revision:

Maxime and Gregory can validate Clara's timesheets,
Maxime and Gregory can edit fields of Clara's validated
timesheets.

opw-3613473

closes odoo/enterprise#59078

X-original-commit: 11d62bcedde2789c5b997205f5570a18eec9ffcb
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>