[IMP] appointment: add form dialog on sharing availabilities
In order to ease the use of custom appointment types,
the user, after selecting custom slots in the calendar view,
can go to "Configure" button. This creates the appointment
type, and before this commit, it redirected to the form view.

In order to stay on this view and keep the flow simple and
smooth, the button now opens a custom dialog form view, simplified
and having custom buttons: "Save and Copy Link" (also closing the
dialog) and "More options" leading to the full form view, as it
was working before.

The user can change the created appointment type through one
of those two form views: the appointment is saved in both cases
when leaving the dialog form with 'Save and Copy Link' or
'More options'.

On closing the dialog using the discard button, the slots will
still appear. This allows easy update, and if one adds or
removes slots, then on opening the dialog again / on copying the
link, the slots will be updated. Once the link is copied in one
of both possible ways, the slots disappear as we consider the
slots properly configured.

In order to have links properly working, even after changing
users in configuration form or full form, we use on creation an
appointment.invite with 'all_assigned_resources' for resource_choice.
This way, any copied link will be redirecting to appropriate users.
A few conditions are added to differentiate this behavior from the
one used when sharing other types of appointments.

Also, to avoid creating useless links, and because this feature
is designed to allow customization when creating the appointment
and invitation on the fly, the same link is passed as props and
used as long as we are on the calendar view and do not use
onClickDiscard.

Task-3805720

closes odoo/enterprise#59017

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>