[IMP] payment*,sale*: render /payment/status entirely from the server
Previously, /payment/status would render a template on the server-side, and a JS widget would render subtemplates on the client-side and insert them in the server-side rendered template.

However, mixing server-side and client-side templates causes issues because the context is not properly communicated between client and server. It also makes the code more complex.

This change assumes that no new information returned by /payment/status is worth re-rendering part of the template. If the transaction moves from a pending ('draft', 'pending') state to a final state, the customer should be redirected to the final route that will display the updated information on the transaction.

task-3340354

closes odoo/enterprise#58822

Related: odoo/upgrade#5829
Related: odoo/documentation#8198
Related: odoo/odoo#149821
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>