[PERF] account_reports_cash_basis: Speedup queries in cash basis
Description

When opening the balance sheet in the cash basis method,
the queries, generated by
`_compute_formula_batch_with_engine_domain`, for the report
 can be slow. The issue is that they filter by date, which
can be slower the more data the
`temp_account_move_line` table contains.

Fix

To speed up these queries, I added a composite index on
the date and a few other fields. Since the index is
added to a temporary table, it will be removed when
the table is dropped.

Benchmark (in 17.0)

| Cash Basis | Before | After |
|------------|--------|-------|
|        1K  |  250ms | 200ms |
|       80K  |    22s |    4s |
|      180K  |    30s |    4s |
|      280K  |    54s |    4s |

(In this case, the 4 seconds come mainly from
the SELECT for payment_table).

Reference

opw-3782652

closes odoo/enterprise#59944

X-original-commit: 026655acb745984ce6cfce61f087e0347f863fd7
Signed-off-by: Aurélien Marie (auma) <auma@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>