[FIX] sale_ebay: sync orders with the most recent product
Steps to reproduce the problem:
- create a product and list it on eBay
- uncheck the sell on eBay setting for the product. Don't archive it
- create a second product and list it to the existing listing in eBay
- sync an order with that product
==> First created product is shown in the sale order

The eBay_id stays on the product, even after unchecking the setting.
This seems to be voluntary as when relisting a product, after some
time without selling it on eBay for instance, this id will then be
used. We now take the first product that is checked as used in eBay.

opw-3503924

closes odoo/enterprise#59065

X-original-commit: aca9a5bfbdbb727f30323e4ec5e2d621f51dbd75
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>