[IMP] spreadsheet_edition: allow to customise groups and measures
This commit allows to customise the groups and measures of a pivot table.
It's possible to edit the aggregate function of a measure and the order
and granularity of a group. The granularity is only available for date
groups.

The groups and measures are editable from the side panel.

closes odoo/enterprise#58957

Task: 3813563
Related: odoo/upgrade#5834
Related: odoo/odoo#158253
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Co-authored-by: Lucas Lefevre <lul@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>