[IMP] helpdesk: use form view with html expander for helpdesk ticket
Before this commit, the description field displayed in the form view of
`helpdesk.ticket` just has a standard height instead of directly taking
the most available spaces as it is the case in the form view of
`project.task` model.

This commit uses the form view with html expander for the form view of
`helpdesk.ticket` model to have the same behavior then the html field
displayed in the form view. That is, compute the height of the
description field to take the highest height possible.

task-3258533

Part-of: odoo/enterprise#59552
Related: odoo/odoo#159496
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>