[FIX] appointment, *: remove record link before test
*: appointment_hr, test_appointment_full

We must disable the cache for the header when there are record links in
it. When the runbot performs enterprise tests, "Helpdesk" appears in the
header, which includes a record link. This commit removes record like
urls in the menu before doing the performance test to prevent increasing
the query count.

opw-3694651
opw-3750925
opw-3781668

closes odoo/enterprise#59321

X-original-commit: b61e676b59363d0529c4e9c9db3df3d97826c04e
Related: odoo/odoo#159006
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Guillaume Dieleman (gdi) <gdi@odoo.com>