[FIX] appointment, *: remove record link before test
*: appointment_hr, test_appointment_full

We must disable the cache for the header when there are record links in
it. When the runbot performs enterprise tests, "Helpdesk" appears in the
header, which includes a record link. This commit removes record like
urls in the menu before doing the performance test to prevent increasing
the query count.

opw-3694651
opw-3750925
opw-3781668

closes odoo/enterprise#59467

X-original-commit: 5e6c02eee897510fe2497b1bf8c7383d01abcc1b
Related: odoo/odoo#159361
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Guillaume Dieleman (gdi) <gdi@odoo.com>