[FIX] account_reports : rearranged account_report_column for Journal Report.
This commit resolves an issue with Journal Report, where any columns positioned after 'additional_col_1' or 'additional_col_2' were not being displayed.
Currently, in every database that migrated to version 17.0, a new record 'invoice_date' is inserted into the account_report_column table and 'Invoice Date' column is not visible in the journal report.

This commit fixes the issue by rearranging the columns when the column lines are being loaded, by pushing 'additional_col_1' and 'additional_col_2' in end of the list.

closes odoo/enterprise#58858

Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>