[FIX] web_studio: remove approvals on button
On the form view editor, click on a button that has studio_approval="True" and some rules behind it
Before this commit, unchecking the "Set approval rules" checkbox in the form view editor
did not disable the rule itself, causing the expected workflow to trigger the rule, even if the button
should not be enforcing them.
This was because of commit 3db107ae481a8da85aa356c729bae5fc246e655b, which introduced server-side
enforcement of studio approval rules.

After this commit, we archive rules that were set on the button, so the expected outside studio workflow works

opw-3810516

closes odoo/enterprise#58982

X-original-commit: 0fa1c253e93a1fcc96801a0953c37a3f86c753a0
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>