[IMP] l10n_us_check_printing: blank checks reports
For the United States, Odoo currently offers the option to print checks
using a "pre-printed" format that matches a specific
supplier (Checkdepot). We don't cover the use case where checks are
printed in-house without a pre-populated page with some static
information. Also, if a company would like to change or edit some
information from the pre-filled checks, it is impossible. 

We want to offer users the option to print "blank checks" by having a
prepopulated template in Odoo that will adjust to the company's
information.

This commit adds 3 additional check layout to print:

- print_blank_check_top
- print_blank_check_middle
- print_blank_check_bottom

Where each of them are designed meticulously to position the required
fields in blank checks, according to Checkdepot's layout.

closes odoo/enterprise#59166

Task-id: 3719584
Signed-off-by: Josse Colpaert <jco@odoo.com>