[FIX] web_studio: make sure auto-save feature is working in studio
In studio, click on the "Automation" menu item.
Create, or edit, one base.automation record, and modify some fields.
Hit the button "Close" to get out of Studio.

Before this commit, the base automation record was not saved.
This was because the studio's action service did not receive the clear uncommitted changes
from the main one.

After this commit, the record is saved when we leave studio.

Note that the ViewEditor mode should not be affected as we operate under a clean owl environment
in StudioView.

closes odoo/enterprise#59196

X-original-commit: aeefa539c69b61c670b61fa99c197306803ee02e
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>