[IMP] stock_barcode(_*): split moves
When the user quits a unfinished picking, the reservation on the move
lines are lost. To avoid that, uncompleted moves will be split in such
case. For example:

- Assume we have a move of 0/15;
- The user picks 4 qty (4/15) then exit the picking:

=> Before this commit, we're left with only one `stock.move.line` with 4
qty and already picked (not knowing what its reservation was since this
information is on the `stock.move` and not on the `stock.move.line`.)

=> After this commit, the line's move will be split and we'll be left
with two `stock.move` and thus `stock.move.line`:
  - One with 4 qty, already picked (4/4);
  - One with 11 qty, waiting to be picked (0/11);

So, the user will always know what the remaining qty they still have to
pick.  Same is true for manufacturing orders since they use
`stock.move` and `stock.move.line` too.

task-3821067

closes odoo/enterprise#61761

X-original-commit: 0ea943dfcd86361cd23380bebcef98e170c70cda
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>