[IMP] web: domain selector: add (not) any operator
In this commit, for the resModel documents.document, we don't want to
have the toggle button included archives. We had to create a
getShowArchivedCheckBox method in domain_selector.js to be able to
patch it in
enterprise/documents/static/src/web/documents_domain_selector.js
so that this functionality still works.

task~3599274

closes odoo/enterprise#58871

Related: odoo/odoo#143949
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>