[IMP] appointment: prevent multiple form submits
Purpose
=======
Prevent multiple submits of the attendee form when the user
spam clicks on the "Confirm Appointment" button.

Specifications
==============
Spam clicking the submit button is submitting multiple times the
form which redirects the user to the "Date & Time" step with a
red banner indicating that another user already has reserved this
time slot a few seconds ago.

Adding a loading effect on the confirmation button to disable it
after a single click and inform the user that the form submission
is correctly been processed.

Task-3815493

closes odoo/enterprise#58865

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>