[IMP] approvals: changed default filter on products
Before this commit the default filter on the product form of approval was
'Can be Sold'. But in approvals, majority users uses purchase related products.

This commit adds the default filter of 'Can be Purchased' on the products and
products variants form of approval. Same default filter is also applied on
more product list for selecting products for approval request.

task-3633895

closes odoo/enterprise#58996

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>