[IMP] account_reports: Journal Report refactoring
The old Journal Report was made in order to have a structured export of the move lines available, via its xlsx and pdf exports. Because of that, its user interface was heavy (since it needed to reflect what we wanted in the pdf/xlsx), and its computation was both slow and complex.

Therefore, we now make the choice to simplify a lot what is shown in the UI by replacing the core of the report by a custom engine (which means the user can now use custom groupby values for this part, so it's more flexible). Only KPIs on the journals used in the periods, and the taxes used within them are shown, not the details of all the move lines. Of course, it is still possible to audit the amounts to see  the details behind them, like on any report, so there is no functional loss (before, unfolding all the journals manually and loading more elements until you found the one you needed was useless anyway => people audited the lines to see their details directly within a tree view and used the search feature from there).

In the meantime, we keep pdf and xlsx exports as complete as before (so, they do show way more information than the UI, and keep a full listing of all the move lines). For XSLX, we now ensure a new page is used for the rendering of every distinct journal. For PDF, we improve the style of the report, for better readability.

task-3698843

closes odoo/enterprise#59778

Related: odoo/upgrade#5897
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>