[FIX] planning: be able to see profile button correctly
Currently, when the user and the target both are in multiple companies,
the profile button cannot be displayed correctly. Since the employee_id
uses `('company_id', '=', self.env.company.id)` rather than `in`.

In community branch we fixed it by checking employee_ids directly and if it
is found, the profile button will be displayed correctly.

In order to accommodate that change, we needed to filter out employee_ids
and add employee_id in fields for patchAvatarCardResourcePopover,
as resource.resource has no field employee_ids.

closes odoo/enterprise#59295

X-original-commit: e3325a3eff631a9ad1e62a9424694feace2f3648
Related: odoo/odoo#158935
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>