[IMP] knowledge: add images as possible anchors
This commit adds the image elements as possible anchors for the comments.
They were until now not considered as it required some last minute adjustments
at the time.
This commit adds those adjustments as well and could be extended to other types
of elements in the future if necessary.

task-3672422

Part-of: odoo/enterprise#54716