[IMP] point_of_sale: use form view for partner editor
In the corresponding comunity commit we replace the custom form view for
editing `res.partner`records with the default `view` component from `web`.

This allows us to remove a very large quantity of code, while providing
for a smoother and more uniform user experience.

In this commit we remove the overrides that we no longer need.

closes odoo/enterprise#59108

Task: 3816230
Related: odoo/odoo#157881
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>