[FIX] sale_stock_renting: select all SNs in view
**Current behavior:**
The `rental.schedule` model/view does not have serial number
data for order lines which are constructed while the
`Rental Transfers` config option is active. This prevents the
view from accurately grouping by serial number.

**Expected behavior:**
The schedule view will correctly group the records in this gantt
view.

**Steps to reproduce:**
1. Create a new rental product, ensure it has the 'By Unique
     Serial Number' tracking option

2. Create an on hand instance of this product with a serial
     number

3. In Rental settings, enable the `Rental Transfers` option

3. Create a new rental order for this product, go to the
     `Schedule` view in the Rental app, select Group By Serial
     Number, observe that the new order is listed as having an
     undefined serial number

**Cause of the issue:**
When the Rental Transfers option is enabled, the wizard which
normally tracks rental pickups/returns gets circumvented to
instead get organized via stock. The SQL view for the schedule
model only looks in the tables managed by this wizard to find
lot_ids of products included in rental orders. So when we don't
use the wizard, that information never makes it to the proper
place.

**Fix:**
Union the current view's table result with another query which
will select the data for the excluded records.

opw-3754422

closes odoo/enterprise#58986

Signed-off-by: Vincent Ethan <etvi@odoo.com>