[FIX] documents: avoid bubbling multipage computation exceptions
As seen in support tickets, other errors can be raised by PyPDF2
when trying to compute pages count on (somewhat) corrupted files.

As we are not going to do any kind of handling and must not crash
because of this, we'll simply log and discard any non-exit
exception raised.

Task-3614532

closes odoo/enterprise#61801

X-original-commit: cabd4fd75f29f3a60e65b49d154df8ee88cb6128
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>