[FIX] web_enterprise: fix client action display name
Since [1], it's now possible to set a display name on a client action,
but a fallback to the action's name was also added. The issue with this,
is that the action's name is a very technical one, and we don't want to
display it to the user.

[1] https://github.com/odoo/odoo/commit/3ad4fd65387f60b524e5f786556963ead8ae9dfe

closes odoo/enterprise#59158

X-original-commit: 86c5de45632fe82b447831ef8cce799d8d541ba3
Related: odoo/odoo#158611
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>