[IMP] l10n_ar_edi: update afip error messages.
Steps to reproduce:

1) Go to runbot Odoo 16 enterprise and install "l10n_ar_edi" module (Argentinean Electronic Invoicing).

2) Take position on Argentinian company (AR) (Responsable Inscripto) .

3) Create customer electronic invoice and confirm. If there is a response with 503 error (HTTPError: 503 Server Error. Service Unavailable) while connecting to the webservice then it will be raised the error message and this text "Please report this error to your Odoo provider" (but this text is not suitable because the odoo provider can`t solve the error. The webservice is not available).

Current behavior:

If there is a response with 503 error (HTTPError: 503 Server Error. Service Unavailable) while connecting to the webservice when the user is trying to confirm an electronic customer invoice then it will be raised the error message and this text "Please report this error to your Odoo provider".

Expected behavior:

If there is a response with 503 error (HTTPError: 503 Server Error. Service Unavailable) while connecting to the webservice when the user is trying to confirm an electronic customer invoice then it will be raised the error message and this text 'The AFIP electronic billing webservice is not available. Wait a few minutes for it to reset and try to validate the action again.'.

Task Adhoc: 37771

closes odoo/enterprise#60638

X-original-commit: ddc196537d53f995652ec3b7c05d70b4194b3c0d
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>