[IMP] hr_payroll: make readonly fields in salary attachment form & list view
In salary attachment user can change some fields even after the payslips have been generated
the fields should be readonly when any one of the payslip is done or paid.

so in this commit added a new boolean field and made a compute function to
check the payslip state and made the necessary fields readonly using the boolean field.

task-3794130

closes odoo/enterprise#59100

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>